+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Added registration and assigned facility IDs
 - Implement patient line list CSV generator
 - Add serial numbers in patient line list CSV
+- Add patient line list downloader
 
 ## 2022-10-03-8443